refactor: move static data into separate data files to enhance modularization
