+++ src/data/catchPhrasesData.js
@@ -1,0 +1,10 @@
+let catchPhrasesArray = [
+    'Check-in to comfort, check-out with memories. Your next unforgettable stay awaits,',
+    'Step into luxury, book your stay today – where comfort meets elegance, and every stay is a dream come true!',
+    "Two words, 'endless comfort !'. Embrace the extraordinary at SwiftStay, where every reservation is a journey of delight.",
+    'A world of luxury awaits behind our double doors. Secure your stay and indulge our elegance.',
+    'Welcome to a symphony of comfort and style. Your key to an unforgettable stay is just a click away – book now at SwiftStay.',
+];
+
+
+export default catchPhrasesArray;
