+++ src/data/hotelsData.js
@@ -1,0 +1,72 @@
+import movenpick_cover_photo from '../Media Files/Hotel Cover Pages/movempick_cover_photo.jpg';
+import kempinski_cover_photo from '../Media Files/Hotel Cover Pages/kempinski_hotel_cover.jpg';
+import la_villa_boutique_cover_photo from '../Media Files/Hotel Cover Pages/la_villa_hotel_cover.jpg';
+import villa_monticello_cover_photo from '../Media Files/Hotel Cover Pages/villa_monticello_hotel_cover.jpg';
+import marriott_cover_photo from '../Media Files/Hotel Cover Pages/marriott_hotel_cover.jpg';
+import number_one_cover_photo from '../Media Files/Hotel Cover Pages/oxford_street_hotel_cover.jpg';
+import holiday_inn_cover_photo from '../Media Files/Hotel Cover Pages/holiday_inn_hotel_cover.jpg';
+import labadi_beach_cover_photo from '../Media Files/Hotel Cover Pages/labadi_beach_hotel_cover.webp';
+
+  
+  
+let AllHotelsArray = [
+    {
+        _id: '63b5934412ae5b2e87f8570c',
+        room_number: 'Movenpick Ambassador Hotel, Accra',
+        room_cover_photo_url: movenpick_cover_photo,
+        room_rate: 5348,
+        room_rating: 4,
+    },
+    {
+        _id: '63b33e0f76cb62893faa578e',
+        room_number: 'Kempinski Hotel Gold Coast City, Accra',
+        room_cover_photo_url: kempinski_cover_photo,
+        room_rate: 5085,
+        room_rating: 4,
+    },
+    {
+        _id: '63b5935912ae5b2e87f8570e',
+        room_number: 'La Villa Boutique Hotel, Accra',
+        room_cover_photo_url: la_villa_boutique_cover_photo,
+        room_rate: 3320,
+        room_rating: 3,
+    },
+    {
+        _id: '640729815f7b99d9a9cc8c1b',
+        room_number: 'Holiday Inn Airport Hotel, Accra',
+        room_cover_photo_url: holiday_inn_cover_photo,
+        room_rate: 2954,
+        room_rating: 4,
+    },
+    {
+        _id: '63b5936f12ae5b2e87f85710',
+        room_number: 'Villa Monticello Boutique Hotel, Accra',
+        room_cover_photo_url: villa_monticello_cover_photo,
+        room_rate: 3526,
+        room_rating: 3,
+    },
+    {
+        _id: '63fcb2ce31251dbf79c2caac',
+        room_number: 'Accra Marriott Hotel Airport, Accra',
+        room_cover_photo_url: marriott_cover_photo,
+        room_rate: 7315,
+        room_rating: 5,
+    },
+    {
+        _id: '64072c225f7b99d9a9cc8c1f',
+        room_number: 'Number One Oxford Street Hotel & Suites, Osu',
+        room_cover_photo_url: number_one_cover_photo,
+        room_rate: 3753,
+        room_rating: 4,
+    },
+    {
+        _id: '64072d315f7b99d9a9cc8c23',
+        room_number: 'Labadi Beach Hotel & Resort, Accra',
+        room_cover_photo_url: labadi_beach_cover_photo,
+        room_rate: 4116,
+        room_rating: 4,
+    },
+];
+
+
+export default AllHotelsArray
