+++ src/data/meetingRoomData.js
@@ -1,0 +1,16 @@
+// for meeting room images
+import meeting_room_image1 from '../Media Files/Meeting Room Images/meeting_room_image1.jpg';
+import meeting_room_image2 from '../Media Files/Meeting Room Images/meeting_room_image2.jpg';
+import meeting_room_image3 from '../Media Files/Meeting Room Images/meeting_room_image3.webp';
+import meeting_room_image4 from '../Media Files/Meeting Room Images/meeting_room_image4.jpg';
+
+
+let meeting_room_pictures_array = [
+    { src: meeting_room_image1 },
+    { src: meeting_room_image2 },
+    { src: meeting_room_image3 },
+    { src: meeting_room_image4 },
+];
+
+
+export default meeting_room_pictures_array;
