+++ src/data/restaurantsData.js
@@ -1,0 +1,15 @@
+import restaurant_image1 from '../Media Files/Restaurants Images/restaurant_image1.jpg';
+import restaurant_image2 from '../Media Files/Restaurants Images/restaurant_image2.webp';
+import restaurant_image3 from '../Media Files/Restaurants Images/restaurant_image3.jpg';
+import restaurant_image4 from '../Media Files/Restaurants Images/restaurant_image4.jpg';
+
+
+let restaurants_pictures_array = [
+    { src: restaurant_image1 },
+    { src: restaurant_image2 },
+    { src: restaurant_image3 },
+    { src: restaurant_image4 },
+];
+
+
+export default restaurants_pictures_array;
