+++ src/data/roomsAndSuitesData.js
@@ -1,0 +1,15 @@
+import rooms_and_suites_1 from '../Media Files/Rooms And Suites/rooms_and_suites_1.webp';
+import rooms_and_suites_2 from '../Media Files/Rooms And Suites/rooms_and_suites_2.jpg';
+import rooms_and_suites_3 from '../Media Files/Rooms And Suites/rooms_and_suites_3.jpg';
+import rooms_and_suites_4 from '../Media Files/Rooms And Suites/rooms_and_suites_4.webp';
+
+
+let rooms_and_suites_pictures_array = [
+    { src: rooms_and_suites_1 },
+    { src: rooms_and_suites_2 },
+    { src: rooms_and_suites_3 },
+    { src: rooms_and_suites_4 },
+];
+
+
+export default rooms_and_suites_pictures_array;
